gbif_deep_learning/requirements.txt
# ── Core deep learning ──────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Data / image handling ────────────────────────────────────────────────────
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0

# ── Scientific computing ─────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0

# ── Visualisation ────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Grad-CAM heatmaps ────────────────────────────────────────────────────────
opencv-python-headless>=4.9.0
